+++ accounting/src/sdk/sdk.ts
@@ -76,9 +76,9 @@ export class SDKConfiguration {
     serverDefaults: any;
     language = "typescript";
     openapiDocVersion = "3.0.0";
-    sdkVersion = "0.7.0";
-    genVersion = "2.248.6";
-    userAgent = "speakeasy-sdk/typescript 0.7.0 2.248.6 3.0.0 @speakeasy-sdks/accounting";
+    sdkVersion = "0.7.1";
+    genVersion = "2.262.2";
+    userAgent = "speakeasy-sdk/typescript 0.7.1 2.262.2 3.0.0 @speakeasy-sdks/accounting";
     retryConfig?: utils.RetryConfig;
     public constructor(init?: Partial<SDKConfiguration>) {
         Object.assign(this, init);